src/presentation/controllers/survey/load-survey/load-surveys-controller.ts
import { LoadSurveys } from '@/domain/usecases/load-surveys'
import { noContent, ok, serverError } from '@/presentation/helpers/http/http-helper'
import { Controller, HttpRequest, HttpResponse } from './load-surveys-protocols'

export class LoadSurveyController implements Controller {
  constructor (private readonly loadSurvey: LoadSurveys) { }

  async handle (_httpRequest: HttpRequest): Promise<HttpResponse> {
    try {
      const surveys = await this.loadSurvey.load()
      if (!surveys) {
        return noContent()
      }
      return ok(surveys)
    } catch (error) {
      return serverError(error)
    }
  }
}
